+++ scout/parse/variant/deleteriousness.py
@@ -1,6 +1,14 @@
 
 def parse_cadd(variant, transcripts):
-    """Check if the cadd phred score is annotated"""
+    """Check if the cadd phred score is annotated
+    
+    Args:
+        variant(cyvcf2.Variant)
+        transcripts(list(dict))
+    
+    Returns:
+        cadd(int)
+    """
     cadd = 0
     cadd_keys = ['CADD', 'CADD_PHRED']
     for key in cadd_keys: